requirements.txt
tls-client==0.2.2
beautifulsoup4==4.12.3
colorama==0.4.6
aiohttp==3.9.3
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
typing-extensions==4.9.0
psutil>=5.9.0
speedtest-cli==2.1.3 